Add subplot values which can be stored in dictionary
Change from using the plt object directly to using a subplot which will
be able to be put into a list for later usage.

The list should be a dictionary of column name to plot for column.

Update the "__main__" usage to properly take the list and pass it to
save images.