Fix -Wignored-qualifiers warning
Don't return a "const" POD; there is no such thing. (This is somewhat
ugly as the code in question is actually a template, but appears to only
be used in one instance on an enumeration type, and looks like it is
anyway intended to be used on enumeration types. Besides, even for class
types, return a const value is of somewhat dubious utility.)